Move the MVarDefault impl into the io-sim package
It's actually not a suitable general purpose implementation, but it is
ok for the simulation.
